@@ -10,27 +10,28 @@
   },
   "dependencies": {
     "@chakra-ui/icons": "^1.1.7",
-    "@chakra-ui/react": "^1.8.7",
+    "@chakra-ui/react": "^1.8.8",
     "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
     "@game-watch/shared": "0.0.1",
     "axios": "^0.26.1",
-    "dayjs": "^1.11.0",
+    "chakra-react-select": "^3.3.1",
+    "dayjs": "^1.11.1",
     "framer-motion": "^5.3.3",
-    "next": "^12.1.4",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-hook-form": "^7.29.0",
+    "next": "^12.1.6",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-hook-form": "^7.30.0",
     "react-masonry-css": "^1.0.16",
-    "sharp": "^0.30.3",
+    "sharp": "^0.30.4",
     "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@types/node": "^17.0.23",
-    "@types/react": "17.0.43",
+    "@types/node": "^17.0.31",
+    "@types/react": "^17.0.43",
     "@types/uuid": "^8.3.4",
-    "eslint": "8.13.0",
-    "eslint-config-next": "12.1.4",
-    "typescript": "^4.6.3"
+    "eslint": "^8.14.0",
+    "eslint-config-next": "^12.1.6",
+    "typescript": "^4.6.4"
   }
 }
