@@ -5,13 +5,14 @@
     "main": "dist/main.js",
     "types": "dist/main.d.ts",
     "scripts": {
-        "build": "tsc -b ."
+        "build": "tsc -b .",
+        "lint": "eslint \"src/**/*.ts\" --fix"
     },
     "devDependencies": {
-        "eslint": "8.13.0",
-        "typescript": "^4.6.3"
+        "eslint": "^8.14.0",
+        "typescript": "^4.6.4"
     },
     "dependencies": {
-        "puppeteer": "^13.5.2"
+        "puppeteer": "^13.7.0"
     }
 }
