@@ -5,17 +5,19 @@
     "main": "dist/main.js",
     "types": "dist/main.d.ts",
     "scripts": {
-        "build": "tsc -b ."
+        "build": "tsc -b .",
+        "lint": "eslint \"src/**/*.ts\" --fix",
+        "mikro-orm": "mikro-orm"
     },
     "dependencies": {
         "@game-watch/shared": "0.0.1",
-        "@mikro-orm/core": "^5.1.1",
-        "@mikro-orm/migrations": "^5.1.1",
+        "@mikro-orm/core": "^5.1.3",
+        "@mikro-orm/migrations": "^5.1.3",
         "@mikro-orm/nestjs": "^5.0.1",
-        "@mikro-orm/postgresql": "^5.1.1",
+        "@mikro-orm/postgresql": "^5.1.3",
         "@mikro-orm/sql-highlighter": "^1.0.1",
         "dotenv": "^16.0.0",
-        "fp-ts": "^2.11.9",
+        "fp-ts": "^2.12.1",
         "io-ts": "^2.2.16",
         "io-ts-reporters": "^2.0.1",
         "io-ts-types": "^0.5.16",
@@ -24,13 +26,13 @@
         "uuid": "^8.3.2"
     },
     "devDependencies": {
-        "@mikro-orm/cli": "^5.1.1",
-        "@types/node": "^17.0.23",
+        "@mikro-orm/cli": "^5.1.3",
+        "@types/node": "^17.0.31",
         "@types/uuid": "^8.3.4",
-        "eslint": "8.13.0",
+        "eslint": "^8.14.0",
         "ts-node": "^10.7.0",
         "tsconfig-paths": "^3.14.1",
-        "typescript": "^4.6.3"
+        "typescript": "^4.6.4"
     },
     "mikro-orm": {
         "useTsNode": true,
