@@ -5,20 +5,23 @@
     "main": "dist/main.js",
     "types": "dist/main.d.ts",
     "scripts": {
-        "build": "tsc -b ."
+        "build": "tsc -b .",
+        "lint": "eslint \"src/**/*.ts\" --fix"
     },
     "devDependencies": {
-        "eslint": "8.13.0",
-        "typescript": "^4.6.3"
+        "eslint": "^8.14.0",
+        "typescript": "^4.6.4"
     },
     "dependencies": {
-        "fp-ts": "^2.11.9",
+        "@game-watch/shared": "0.0.1",
+        "@sentry/node": "^6.19.7",
+        "fp-ts": "^2.12.1",
         "io-ts": "^2.2.16",
         "io-ts-reporters": "^2.0.1",
         "io-ts-types": "^0.5.16",
         "monocle-ts": "^2.3.13",
         "newtype-ts": "^0.3.5",
-        "pino": "^7.10.0",
-        "pino-pretty": "^7.6.0"
+        "pino": "^7.11.0",
+        "pino-pretty": "^7.6.1"
     }
 }
