@@ -5,11 +5,12 @@
     "main": "dist/main.js",
     "types": "dist/main.d.ts",
     "scripts": {
-        "build": "tsc -b ."
+        "build": "tsc -b .",
+        "lint": "eslint \"src/**/*.ts\" --fix"
     },
     "devDependencies": {
-        "eslint": "8.13.0",
-        "typescript": "^4.6.3"
+        "eslint": "^8.14.0",
+        "typescript": "^4.6.4"
     },
     "dependencies": {
         "class-validator": "^0.13.1"
