@@ -9,7 +9,7 @@
         "build": "tsc -b .",
         "start:dev": "ts-node-dev src/main.ts",
         "start:prod": "node dist/main.js",
-        "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix"
+        "lint": "eslint \"src/**/*.ts\" --fix"
     },
     "dependencies": {
         "@game-watch/browser": "0.0.1",
@@ -17,35 +17,36 @@
         "@game-watch/queue": "0.0.1",
         "@game-watch/service": "0.0.1",
         "@game-watch/shared": "0.0.1",
-        "@mikro-orm/core": "^5.1.1",
-        "@mikro-orm/migrations": "^5.1.1",
+        "@mikro-orm/core": "^5.1.3",
+        "@mikro-orm/migrations": "^5.1.3",
         "@mikro-orm/nestjs": "^5.0.1",
-        "@mikro-orm/postgresql": "^5.1.1",
-        "@sentry/node": "^6.19.3",
-        "@sentry/tracing": "^6.19.6",
+        "@mikro-orm/postgresql": "^5.1.3",
+        "@sentry/node": "^6.19.7",
+        "@sentry/tracing": "^6.19.7",
         "axios": "^0.26.1",
-        "bullmq": "^1.78.2",
+        "bullmq": "^1.81.4",
         "cheerio": "^1.0.0-rc.10",
-        "dayjs": "^1.11.0",
-        "fp-ts": "^2.11.9",
+        "dayjs": "^1.11.1",
+        "dotenv": "^16.0.0",
+        "fp-ts": "^2.12.1",
         "io-ts": "^2.2.16",
         "io-ts-types": "^0.5.16",
         "ioredis": "^4.28.2",
         "monocle-ts": "^2.3.13",
         "newtype-ts": "^0.3.5",
-        "p-retry": "^4.6.1",
+        "p-retry": "^4.6.2",
         "reflect-metadata": "^0.1.13"
     },
     "devDependencies": {
         "@types/ioredis": "^4.28.10",
-        "@types/node": "^17.0.23",
-        "@typescript-eslint/eslint-plugin": "^5.17.0",
-        "@typescript-eslint/parser": "^5.18.0",
-        "eslint": "^8.13.0",
+        "@types/node": "^17.0.31",
+        "@typescript-eslint/eslint-plugin": "^5.22.0",
+        "@typescript-eslint/parser": "^5.22.0",
+        "eslint": "^8.14.0",
         "eslint-plugin-simple-import-sort": "^7.0.0",
         "prettier": "^2.6.2",
         "ts-node-dev": "^1.1.8",
         "tsconfig-paths": "^3.14.1",
-        "typescript": "^4.6.3"
+        "typescript": "^4.6.4"
     }
 }
