@@ -11,8 +11,8 @@
     "start": "nest start",
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
-    "start:prod": "node dist/main",
-    "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
+    "start:prod": "node dist/src&main",
+    "lint": "eslint \"src/**/*.ts\" --fix",
     "create-crons-for-games": "ts-node ./bin/create-crons-for-existing-games",
     "create-crons-for-info-sources": "ts-node ./bin/create-crons-for-existing-info-sources"
   },
@@ -21,28 +21,29 @@
     "@game-watch/queue": "0.0.1",
     "@game-watch/service": "0.0.1",
     "@game-watch/shared": "0.0.1",
-    "@mikro-orm/core": "^5.1.1",
-    "@mikro-orm/migrations": "^5.1.1",
+    "@mikro-orm/core": "^5.1.3",
+    "@mikro-orm/knex": "^5.1.3",
+    "@mikro-orm/migrations": "^5.1.3",
     "@mikro-orm/nestjs": "^5.0.1",
-    "@mikro-orm/postgresql": "^5.1.1",
+    "@mikro-orm/postgresql": "^5.1.3",
     "@nestjs/bullmq": "^0.0.4",
-    "@nestjs/common": "^8.4.3",
+    "@nestjs/common": "^8.4.4",
     "@nestjs/config": "^2.0.0",
     "@nestjs/core": "^8.4.4",
     "@nestjs/jwt": "^8.0.0",
     "@nestjs/passport": "^8.2.1",
     "@nestjs/platform-express": "^8.4.4",
     "@nestjs/throttler": "^2.0.1",
-    "@sentry/node": "^6.19.3",
-    "@sentry/tracing": "^6.19.6",
+    "@sentry/node": "^6.19.7",
+    "@sentry/tracing": "^6.19.7",
     "bcrypt": "^5.0.1",
-    "bullmq": "^1.78.2",
+    "bullmq": "^1.81.4",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.13.1",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.6",
     "dotenv": "^16.0.0",
-    "fp-ts": "^2.11.9",
+    "fp-ts": "^2.12.1",
     "io-ts": "^2.2.16",
     "io-ts-types": "^0.5.16",
     "monocle-ts": "^2.3.13",
@@ -53,30 +54,34 @@
     "passport-jwt": "^4.0.0",
     "passport-local": "^1.0.0",
     "pino-http": "^6.5.0",
-    "pino-pretty": "^7.6.0",
+    "pino-pretty": "^7.6.1",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^7.5.5",
-    "uuid": "^8.3.2",
-    "typescript": "^4.6.3",
-    "ts-node": "^10.7.0"
+    "ts-node": "^10.7.0",
+    "typescript": "^4.6.4",
+    "uuid": "^8.3.2"
   },
   "devDependencies": {
-    "@nestjs/cli": "^8.2.4",
-    "@nestjs/schematics": "^8.0.0",
+    "@nestjs/cli": "^8.2.5",
+    "@nestjs/schematics": "^8.0.11",
     "@types/bcrypt": "^5.0.0",
     "@types/compression": "^1.7.2",
-    "@types/cookie-parser": "^1.4.2",
+    "@types/cookie-parser": "^1.4.3",
     "@types/express": "^4.17.13",
-    "@types/node": "^17.0.23",
+    "@types/node": "^17.0.31",
     "@types/passport-jwt": "^3.0.6",
     "@types/passport-local": "^1.0.34",
     "@types/uuid": "^8.3.4",
-    "@typescript-eslint/parser": "^5.18.0",
-    "eslint": "^8.13.0",
+    "@typescript-eslint/parser": "^5.22.0",
+    "eslint": "^8.14.0",
     "eslint-plugin-simple-import-sort": "^7.0.0",
     "prettier": "^2.6.2",
-    "ts-loader": "^9.2.8",
+    "ts-loader": "^9.3.0",
     "tsconfig-paths": "^3.14.1"
+  },
+  "peerDependencies": {
+    "reflect-metadata": "^0.1.12",
+    "rxjs": "^7.1.0"
   }
 }
