@@ -1,63 +1,82 @@
-import { Box } from "@chakra-ui/layout"
 import {
+    Box,
     Stat,
     StatLabel,
     StatNumber,
     Text,
-} from "@chakra-ui/react"
-import React, { useCallback, useMemo } from "react"
-import dayjs from "dayjs"
-import { InfoSourceWrapper } from "./InfoSourceWrapper"
-import { StoreGameData } from "@game-watch/shared"
+} from "@chakra-ui/react";
+import { Country, StoreGameData } from "@game-watch/shared";
+import dayjs from "dayjs";
+import React, { useCallback, useMemo } from "react";
 
-var customParseFormat = require('dayjs/plugin/customParseFormat')
-dayjs.extend(customParseFormat)
+import { useUserContext } from "../../providers/UserProvider";
+import { InfoSourceWrapper } from "./InfoSourceWrapper";
+
+// eslint-disable-next-line @typescript-eslint/no-var-requires
+const customParseFormat = require('dayjs/plugin/customParseFormat');
+dayjs.extend(customParseFormat);
 
 const ReleaseDate: React.FC<{ date?: Date }> = ({ date }) => {
-    const formattedDate = useMemo(() => date ? dayjs(date).format("DD MMM, YYYY") : "TBD", [date])
+    const formattedDate = useMemo(() => date ? dayjs(date).format("DD MMM, YYYY") : "TBD", [date]);
 
     return (
         <Stat>
             <StatLabel>Release Date</StatLabel>
             <StatNumber fontSize="1rem">{formattedDate}</StatNumber>
         </Stat>
-    )
-}
+    );
+};
 
-const Price: React.FC<{ price?: number, initial?: number }> = ({ price, initial }) => {
+
+const countryUnitMap: Record<Country, string> = {
+    "DE": "€",
+    "US": "$"
+};
+
+const Price: React.FC<{ price?: number, initial?: number, userCountry: Country }> = ({ price, initial, userCountry }) => {
     const formatPrice = useCallback((price?: number) => {
         if (price === undefined) {
-            return "TBA"
+            return "TBA";
         }
 
         if (price === 0) {
-            return "Free"
+            return "Free";
         }
 
-        return `${price}€`
-    }, [])
+        if (countryUnitMap[userCountry] === "$") {
+            return `$${price}`;
+        }
 
-    const hasDiscount = useMemo(() => price && initial && price !== initial, [price, initial])
-    const parsedPrice = useMemo(() => formatPrice(price), [formatPrice, price])
-    const parsedInitial = useMemo(() => formatPrice(initial), [formatPrice, initial])
+        return `${price}${countryUnitMap[userCountry]}`;
+    }, [userCountry]);
+
+    const hasDiscount = useMemo(() => price && initial && price !== initial, [price, initial]);
+    const parsedPrice = useMemo(() => formatPrice(price), [formatPrice, price]);
+    const parsedInitial = useMemo(() => formatPrice(initial), [formatPrice, initial]);
 
     return (
         <Stat>
             <StatLabel>Price</StatLabel>
             <StatNumber fontSize="1rem">{hasDiscount ? <Text as="s">{parsedInitial}</Text> : null} {parsedPrice}</StatNumber>
         </Stat>
-    )
-}
+    );
+};
 
 export const StoreInfoSource: React.FC<{ data: StoreGameData | null, }> = ({ data }) => {
+    const { user } = useUserContext();
+
     return (
         <InfoSourceWrapper>
             <Box flex="1">
                 <ReleaseDate date={data?.releaseDate} />
             </Box>
             <Box flex="1">
-                <Price price={data?.priceInformation?.final} initial={data?.priceInformation?.initial} />
+                <Price
+                    price={data?.priceInformation?.final}
+                    initial={data?.priceInformation?.initial}
+                    userCountry={user.country}
+                />
             </Box>
         </InfoSourceWrapper>
-    )
-}
+    );
+};
