@@ -1,11 +1,6 @@
-import { Input } from "@chakra-ui/input";
-import {
-    Button,
-    useColorModeValue,
-    useDisclosure,
-} from "@chakra-ui/react";
-import { Flex } from "@chakra-ui/layout";
-import React, { useState, useCallback, useRef, useEffect, useMemo } from "react";
+import { Button, Flex, Input, useColorModeValue, useDisclosure } from "@chakra-ui/react";
+import React, { KeyboardEvent,useCallback, useEffect, useMemo, useRef, useState } from "react";
+
 import { GameProvider } from "../providers/GameProvider";
 import { useGamesContext } from "../providers/GamesProvider";
 import { useAction } from "../util/useAction";
@@ -24,23 +19,23 @@ export const AddGameInput: React.FC = () => {
             setGameId(game.id);
             onOpen();
         }
-    })
+    });
 
     // Focus initially
-    const inputRef = useRef<HTMLInputElement | null>(null)
-    useEffect(() => { inputRef.current && inputRef.current.focus() }, []);
+    const inputRef = useRef<HTMLInputElement | null>(null);
+    useEffect(() => { inputRef.current && inputRef.current.focus(); }, []);
 
-    const onNameKeyPress = useCallback(async ({ key }) => {
+    const onNameKeyPress = useCallback(async ({ key }: KeyboardEvent<HTMLInputElement>) => {
         if (key === "Enter") {
             await searchGame(name);
         }
     }, [searchGame, name]);
 
     const onCloseModal = useCallback(() => {
-        setGameId(null)
+        setGameId(null);
         setName("");
         onClose();
-    }, [onClose])
+    }, [onClose]);
 
     const currentGame = useMemo(() => games.find(game => game.id === gameId), [games, gameId]);
 
@@ -74,5 +69,5 @@ export const AddGameInput: React.FC = () => {
                 </GameProvider>
             }
         </>
-    )
-}
+    );
+};
