@@ -1,39 +1,39 @@
 import {
-    Text,
-    FormControl,
+    Box,
+    Button,
     chakra,
+    Checkbox,
+    Flex,
+    FormControl,
+    FormErrorMessage,
     FormLabel,
     Input,
-    Box,
-    Flex,
-    Button,
-    Checkbox,
-    FormErrorMessage,
+    Text,
     useToast,
-} from "@chakra-ui/react"
-import React from "react"
-import { useUserContext } from "../../providers/UserProvider"
-import { useAction } from "../../util/useAction"
-
-import { useForm } from "react-hook-form";
+} from "@chakra-ui/react";
 import axios from "axios";
+import React from "react";
+import { useForm } from "react-hook-form";
+
+import { useUserContext } from "../../providers/UserProvider";
+import { useAction } from "../../util/useAction";
 
 type Inputs = {
     username: string,
     password: string,
-};
+}
 
 const Form = chakra("form");
 
 export const RegisterUserForm: React.FC<{ onCancel: () => void }> = ({ onCancel }) => {
     const toast = useToast();
-    const { registerUser } = useUserContext()
+    const { registerUser } = useUserContext();
     const { register, handleSubmit, formState: { errors }, setError } = useForm<Inputs>();
 
     const { loading, execute: onRegister } = useAction(registerUser, {
         onError: error => {
             if (axios.isAxiosError(error) && error.response?.status === 409) {
-                return setError("username", { type: "validate" })
+                return setError("username", { type: "validate" });
             }
 
             toast({
@@ -43,7 +43,7 @@ export const RegisterUserForm: React.FC<{ onCancel: () => void }> = ({ onCancel 
                 position: "top-right",
             });
         }
-    })
+    });
 
     return (
         <Box>
@@ -75,5 +75,5 @@ export const RegisterUserForm: React.FC<{ onCancel: () => void }> = ({ onCancel 
                 </Flex>
             </Form>
         </Box>
-    )
-}
+    );
+};
