@@ -1,7 +1,8 @@
-import React, { useEffect, useState } from "react"
-import { ButtonGroup, Editable, EditableInput, EditablePreview, Flex, IconButton, useEditableControls } from "@chakra-ui/react"
-import { CheckIcon, CloseIcon, EditIcon } from "@chakra-ui/icons"
-import { useGameContext } from "../../providers/GameProvider"
+import { CheckIcon, CloseIcon, EditIcon } from "@chakra-ui/icons";
+import { ButtonGroup, Editable, EditableInput, EditablePreview, Flex, IconButton, useEditableControls } from "@chakra-ui/react";
+import React, { useEffect, useState } from "react";
+
+import { useGameContext } from "../../providers/GameProvider";
 
 const EditableControls: React.FC = () => {
     const {
@@ -9,7 +10,7 @@ const EditableControls: React.FC = () => {
         getSubmitButtonProps,
         getCancelButtonProps,
         getEditButtonProps,
-    } = useEditableControls()
+    } = useEditableControls();
 
     return isEditing ? (
         <ButtonGroup justifyContent="center" size="sm" ml="1rem">
@@ -20,15 +21,15 @@ const EditableControls: React.FC = () => {
             <Flex position="absolute" justifyContent="center" top="auto" bottom="auto" right="0">
                 <IconButton aria-label="edit" size="sm" icon={<EditIcon />} {...getEditButtonProps()} />
             </Flex>
-    )
-}
+    );
+};
 
 export const GameName: React.FC<{ disableEdit: boolean }> = ({ disableEdit }) => {
     const { fullName, changeGameName } = useGameContext();
     const [showControls, setShowControls] = useState(false);
     const [value, setValue] = useState(fullName);
 
-    useEffect(() => { setValue(fullName); }, [fullName])
+    useEffect(() => { setValue(fullName); }, [fullName]);
 
     return (
         <Editable
@@ -49,5 +50,5 @@ export const GameName: React.FC<{ disableEdit: boolean }> = ({ disableEdit }) =>
                 {!disableEdit && showControls && <EditableControls />}
             </Flex>
         </Editable>
-    )
-}
+    );
+};
