@@ -1,15 +1,14 @@
-import Image from 'next/image';
-import { Box, Flex } from "@chakra-ui/layout";
+import { Box, Flex, Skeleton, useColorModeValue } from "@chakra-ui/react";
 import React, { useEffect, useState } from "react";
-import { Skeleton, useColorModeValue } from "@chakra-ui/react";
-import { LoadingSpinner } from "../LoadingSpinner";
+
 import { useGameContext } from "../../providers/GameProvider";
+import { LoadingSpinner } from "../LoadingSpinner";
 
 export const GameThumbnail: React.FC = () => {
     const { loading, game, thumbnailUrl } = useGameContext();
 
     const [imageLoading, setImageLoading] = useState(false);
-    useEffect(() => { setImageLoading(true) }, []);
+    useEffect(() => { setImageLoading(true); }, []);
 
     return (
         <Box position="relative">
@@ -17,12 +16,10 @@ export const GameThumbnail: React.FC = () => {
             <Skeleton isLoaded={thumbnailUrl !== null ? !imageLoading : true}>
                 <Flex position="relative" justify="center" height="215px" bg={useColorModeValue("white", "gray.900")} >
                     {thumbnailUrl &&
-                        <Image
-                        quality={100}
-                            alt=""
-                            priority={true}
-                            layout="fill"
-                            objectFit="cover"
+                        <img
+                            width="460"
+                            style={{ objectFit: "cover" }}
+                        alt=""
                             src={thumbnailUrl}
                             onError={() => setImageLoading(false)}
                             onLoad={() => setImageLoading(false)}
@@ -31,5 +28,5 @@ export const GameThumbnail: React.FC = () => {
                 </Flex>
             </Skeleton>
         </Box>
-    )
-}
+    );
+};
