@@ -1,8 +1,9 @@
-import React from "react"
-import { StoreInfoSource } from "./StoreInfoSource"
-import { MetacriticInfoSource } from "./MetacriticInfoSource"
 import { InfoSourceType, MetacriticData } from "@game-watch/shared";
+import React from "react";
+
 import { useInfoSourceContext } from "../../providers/InfoSourceProvider";
+import { MetacriticInfoSource } from "./MetacriticInfoSource";
+import { StoreInfoSource } from "./StoreInfoSource";
 
 export const InfoSource: React.FC = () => {
     const { source } = useInfoSourceContext();
@@ -12,8 +13,8 @@ export const InfoSource: React.FC = () => {
         case InfoSourceType.Switch:
         case InfoSourceType.PsStore:
         case InfoSourceType.Epic:
-            return <StoreInfoSource data={source.data} />
+            return <StoreInfoSource data={source.data} />;
         case InfoSourceType.Metacritic:
             return <MetacriticInfoSource data={source.data as MetacriticData} />;
     }
-}
+};
