@@ -1,12 +1,14 @@
-import React from "react"
-import { Flex, Box } from "@chakra-ui/layout"
 import {
+    Box,
+    Flex,
     Text,
     Tooltip
 } from "@chakra-ui/react";
+import React from "react";
+
+import { useInfoSourceContext } from "../../providers/InfoSourceProvider";
 import { LoadingSpinner } from "../LoadingSpinner";
 import { InfoSourceOptions } from "./InfoSourceOptions";
-import { useInfoSourceContext } from "../../providers/InfoSourceProvider";
 import { SourceTypeLogo } from "./SourceTypeLogo";
 
 export const InfoSourceWrapper: React.FC = ({ children }) => {
@@ -31,6 +33,6 @@ export const InfoSourceWrapper: React.FC = ({ children }) => {
             )}
             <Box><InfoSourceOptions /></Box>
         </Flex>
-    )
-}
+    );
+};
 
