@@ -1,10 +1,12 @@
-import '../styles/globals.css'
-import type { AppProps } from 'next/app'
-import { ChakraProvider } from "@chakra-ui/react"
+import '../styles/globals.css';
+
+import { ChakraProvider } from "@chakra-ui/react";
+import type { AppProps } from 'next/app';
+
 import Layout from "../components/Layout";
-import theme from '../theme';
 import { NotificationProvider } from '../providers/NotificationProvider';
 import { UserProvider } from '../providers/UserProvider';
+import theme from '../theme';
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
@@ -19,4 +21,4 @@ function MyApp({ Component, pageProps }: AppProps) {
     </ChakraProvider>
   );
 }
-export default MyApp
+export default MyApp;
