@@ -1,6 +1,7 @@
-import { ColorModeScript } from "@chakra-ui/react"
-import NextDocument, { Html, Head, Main, NextScript } from "next/document"
-import theme from "../theme"
+import { ColorModeScript } from "@chakra-ui/react";
+import NextDocument, { Head, Html, Main, NextScript } from "next/document";
+
+import theme from "../theme";
 
 export default class Document extends NextDocument {
   render() {
@@ -19,6 +20,6 @@ export default class Document extends NextDocument {
           <NextScript />
         </body>
       </Html>
-    )
+    );
   }
 }
