@@ -1,9 +1,9 @@
-import React, { useCallback, useContext, useEffect, useMemo, useState } from "react";
-import { useHttp } from "../util/useHttp";
-import { AxiosResponse } from "axios";
 import { CreateInfoSourceDto, GameDto, InfoSourceDto, InfoSourceType, TagDto } from "@game-watch/shared";
+import { AxiosResponse } from "axios";
+import React, { useCallback, useContext, useEffect, useMemo, useState } from "react";
 
-// TODO: Let users select the priority / image
+import { useHttp } from "../util/useHttp";
+
 export const INFO_SOURCE_PRIORITY = [
     InfoSourceType.PsStore,
     InfoSourceType.Steam,
@@ -39,7 +39,7 @@ const retrieveDataFromInfoSources = (infoSources: InfoSourceDto[], key: string):
     }
 
     return null;
-}
+};
 
 export interface GameCtx {
     game: GameDto
@@ -55,6 +55,7 @@ export interface GameCtx {
     changeGameName: (name: string) => Promise<void>
     deleteGame: () => Promise<void>
     setGameInfoSource: (infoSource: InfoSourceDto) => void
+    updateGameInfoSource: (infoSource: InfoSourceDto) => void
     removeGameInfoSource: (id: string) => void
     addTagToGame: (tag: TagDto) => Promise<void>
     removeTagFromGame: (tag: TagDto) => Promise<void>
@@ -72,6 +73,7 @@ export function useGameContext() {
 }
 
 export const GameProvider: React.FC<{
+    children: React.ReactChild,
     game: GameDto,
     setGame: (id: string, cb: ((current: GameDto) => GameDto) | GameDto) => void
     removeGame: (id: string) => void
@@ -94,6 +96,7 @@ export const GameProvider: React.FC<{
             return;
         }
         setPolling(true);
+
         (async () => {
             await withRequest(async http => {
                 do {
@@ -108,7 +111,7 @@ export const GameProvider: React.FC<{
                     } finally {
                         await new Promise(resolve => setTimeout(resolve, 1000));
                     }
-                } while (true)
+                } while (true);
             });
             setPolling(false);
         }
@@ -159,6 +162,19 @@ export const GameProvider: React.FC<{
     const setGameInfoSource = useCallback((newInfoSource: InfoSourceDto) => {
         setGame(game.id, curr => {
             curr.infoSources = [...curr.infoSources.filter(({ id }) => newInfoSource.id !== id), newInfoSource];
+            return curr;
+        });
+    }, [setGame, game.id]);
+
+    const updateGameInfoSource = useCallback((newInfoSource: InfoSourceDto) => {
+        setGame(game.id, curr => {
+            const filteredInfoSources = curr.infoSources.filter(({ id }) => newInfoSource.id !== id);
+            // Make sure sources disabled in the mean time are not re-added
+            if (filteredInfoSources.length === curr.infoSources.length) {
+                return curr;
+            }
+
+            curr.infoSources = [...filteredInfoSources, newInfoSource];
             return curr;
         });
     }, [setGame, game.id]);
@@ -284,6 +300,7 @@ export const GameProvider: React.FC<{
         changeGameName,
         deleteGame,
         setGameInfoSource,
+        updateGameInfoSource,
         removeGameInfoSource,
         addTagToGame,
         removeTagFromGame,
@@ -302,6 +319,7 @@ export const GameProvider: React.FC<{
         changeGameName,
         deleteGame,
         setGameInfoSource,
+        updateGameInfoSource,
         removeGameInfoSource,
         addTagToGame,
         removeTagFromGame,
@@ -312,5 +330,5 @@ export const GameProvider: React.FC<{
         <GameContext.Provider value={contextValue}>
             {children}
         </GameContext.Provider>
-    )
-}
+        );
+    };
