@@ -1,30 +1,3 @@
 module.exports = {
-  parser: '@typescript-eslint/parser',
-  parserOptions: {
-    project: 'tsconfig.json',
-    sourceType: 'module',
-  },
-  plugins: [
-    '@typescript-eslint/eslint-plugin',
-    'simple-import-sort'
-  ],
-  extends: [
-    'plugin:@typescript-eslint/recommended'
-  ],
-  root: true,
-  env: {
-    node: true,
-    jest: true,
-  },
-  ignorePatterns: ['.eslintrc.js'],
-  rules: {
-    '@typescript-eslint/interface-name-prefix': 'off',
-    '@typescript-eslint/explicit-function-return-type': 'off',
-    '@typescript-eslint/explicit-module-boundary-types': 'off',
-    '@typescript-eslint/no-explicit-any': 'off',
-    '@typescript-eslint/no-empty-function': 'off',
-    '@typescript-eslint/no-inferrable-types': 'off',
-    '@typescript-eslint/semi': 'error',
-    "simple-import-sort/imports": "error"
-  },
+  extends: "../.eslintrc.js"
 };
