@@ -3,7 +3,7 @@ export const parseCurrencyValue = (value?: string) => {
         return undefined;
     }
 
-    const parsedValue = parseFloat(value.replace("€", "").replace(",", "."));
+    const parsedValue = parseFloat(value.replace(/€|\$/g, "").replace(",", "."));
     if (isNaN(parsedValue)) {
         return undefined;
     }
