@@ -1,3 +1,4 @@
+import { mapCountryCodeToAcceptLanguage, mapCountryCodeToLanguage } from "@game-watch/service";
 import { InfoSourceType } from "@game-watch/shared";
 import axios from "axios";
 
@@ -7,17 +8,18 @@ import { matchingName } from "../util/matching-name";
 export class PsStoreSearcher implements InfoSearcher {
     public type = InfoSourceType.PsStore;
 
-    public async search(search: string, { logger }: InfoSearcherContext) {
+
+    public async search(search: string, { logger, userCountry }: InfoSearcherContext) {
         const response = await axios.get(
             `https://web.np.playstation.com/api/graphql/v1/op`,
             {
                 headers: {
-                    "X-PSN-Store-Locale-Override": "de-DE"
+                    "X-PSN-Store-Locale-Override": mapCountryCodeToAcceptLanguage(userCountry)
                 },
                 params: {
                     variables: {
-                        countryCode: "DE",
-                        languageCode: "de",
+                        countryCode: userCountry,
+                        languageCode: mapCountryCodeToLanguage(userCountry),
                         pageOffset: 0,
                         pageSize: 1,
                         searchTerm: search,
@@ -49,7 +51,7 @@ export class PsStoreSearcher implements InfoSearcher {
         logger.debug(`Found gameId '${gameId}'`);
 
         return {
-            remoteGameId: `https://store.playstation.com/de-de/product/${gameId}`,
+            remoteGameId: `https://store.playstation.com/${mapCountryCodeToAcceptLanguage(userCountry)}/product/${gameId}`,
             remoteGameName: fullName,
         };
     }
